requirements.txt
torch==1.8.0
torchvision==0.9.0
torchaudio==0.8.0
tensorboardx==2.4
augly
opencv-python==4.5.3.56
faiss-cpu==1.7.1.post2
h5py==3.4.0
pandas==1.3.2
typer==0.3.2
matplotlib==3.4.3
faiss-gpu==1.7.1.post2
seaborn
pyyaml 
yacs 
timm
scikit-image
